User to s user

+++ AutoRep/Models/WorkType.cs
@@ -11,7 +11,7 @@ namespace AutoRep.Models
         public int Id { get; set; }
         public string Name { get; set; }//Название работы
         public string Text { get; set; }//Описание работы
-        public int Cost { get; set; }//Стоимость работы
+        public double Cost { get; set; }//Стоимость работы
 
         public enum SortState
         {
